lib/invoice/app/code/requirements.txt
aws-embedded-metrics~=3.2.0
aws-xray-sdk~=2.12.0
boto3~=1.28.59
botocore~=1.31.59
Flask~=2.3.3
gunicorn~=21.2.0
PyJWT~=2.8.0
python-dateutil~=2.8.2
requests~=2.31.0
